+++ samples/c2pa/C2paPlayer/C2paMenu/C2paMenuFunctions.js
@@ -1,4 +1,7 @@
-export let initializeC2PAMenu = function (c2paMenuInstance, videoPlayer) {
+import { C2PAMenu } from './C2paMenu.js';
+//C2PA menu instance
+let c2paMenuInstance = new C2PAMenu();
+export let initializeC2PAMenu = function (videoPlayer) {
     const MenuButton = videojs.getComponent('MenuButton');
     const MenuItem = videojs.getComponent('MenuItem');
 
@@ -62,4 +65,53 @@ export let adjustC2PAMenu = function (c2paMenu , videoElement , c2paMenuHeightOf
 
     menuContent.style.width = `${playerWidth}px`;
     menuContent.style.height = `${playerHeight}px`;
+};
+
+
+//Update the c2pa menu items with the values from the c2pa manifest
+export let updateC2PAMenu = function (c2paStatus, c2paMenu , isMonolithic , videoPlayer , getCompromisedRegions) {
+    //Get all the items in the c2pa menu
+    const c2paItems = c2paMenu.el().querySelectorAll('.vjs-menu-item');
+    const compromisedRegions = getCompromisedRegions(isMonolithic , videoPlayer);
+
+    c2paItems.forEach((c2paItem) => {
+        //Menu items are organized as key/name + value, separated by a delimiter
+        const c2paItemText = c2paItem.innerText;
+        const c2paItemName = c2paItemText.split(
+            c2paMenuInstance.c2paMenuDelimiter()
+        )[0];
+
+        //Based on the plain name of the menu item, we retrieve the key from the c2paMenuInstance
+        //And based on that key, we get the corresponding value from the c2pa manifest
+        const c2paItemKey =
+                c2paMenuInstance.c2paMenuValueToKeyMap(c2paItemName);
+        const c2paItemValue = c2paMenuInstance.c2paItem(
+            c2paItemKey,
+            c2paStatus,
+            compromisedRegions
+        );
+        console.log(
+            '[C2PA] Menu item: ',
+            c2paItemName,
+            c2paItemKey,
+            c2paItemValue
+        );
+
+        if (c2paItemValue != null) {
+            //If the value is not null, we update the menu item text and show it
+            c2paItem.innerText =
+                    c2paItemName +
+                    c2paMenuInstance.c2paMenuDelimiter() +
+                    c2paItemValue;
+            c2paItem.style.display = 'block';
+        } else {
+            //If the value is null, we hide the menu item
+            c2paItem.style.display = 'none';
+        }
+    });
+};
+
+//Hide the c2pa menu
+let hideC2PAMenu = function () {
+    c2paMenu.hide();
 };
