--- samples/microsoft/video_samples/DASH/audio_video_sample/audio_prog_index.m3u8
@@ -1,29 +1,0 @@
-#EXTM3U
-#EXT-X-TARGETDURATION:2
-#EXT-X-VERSION:7
-#EXT-X-MEDIA-SEQUENCE:1
-#EXT-X-PLAYLIST-TYPE:VOD
-#EXT-X-INDEPENDENT-SEGMENTS
-#EXT-X-MAP:URI="provenanceaudiooutput_dashinit.mp4"
-#EXTINF:2,
-provenanceaudiooutput_dash1.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash2.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash3.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash4.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash5.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash6.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash7.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash8.m4s
-#EXTINF:2
-provenanceaudiooutput_dash9.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash10.m4s
-#EXTINF:2,
-provenanceaudiooutput_dash11.m4s
