--- samples/microsoft/video_samples/DASH/audio_video_sample/main.m3u8
@@ -1,5 +1,0 @@
-#EXTM3U
-#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="audio",LANGUAGE="eng",NAME="English",AUTOSELECT=YES,DEFAULT=YES,URI="audio_prog_index.m3u8"
-
-#EXT-X-STREAM-INF:TYPE=VIDEO,AVERAGE-BANDWIDTH=479072,BANDWIDTH=479072,CODECS="avc1.42400D",RESOLUTION=320x240,FRAME-RATE=29.970,CLOSED-CAPTIONS=NONE,AUDIO="audio"
-video_prog_index.m3u8
