--- samples/microsoft/video_samples/DASH/audio_video_sample/manifest.mpd
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<!-- MPD file Generated with GPAC version 0.8.0-rev1-gc1990d5c-master  at 2020-10-14T04:06:54.788Z-->
-<MPD xmlns="urn:mpeg:dash:schema:mpd:2011" minBufferTime="PT1.500S" type="static" mediaPresentationDuration="PT0H0M20.320S" maxSegmentDuration="PT0H0M2.402S" profiles="urn:mpeg:dash:profile:isoff-live:2011">
- <ProgramInformation moreInformationURL="http://gpac.io">
-  <Title>manifest.mpd generated by GPAC</Title>
- </ProgramInformation>
-
- <Period duration="PT0H0M20.320S">
-  <AdaptationSet segmentAlignment="true" lang="eng" startWithSAP="1">
-   <SegmentTemplate media="provenanceaudiooutput_dash$Number$.m4s" initialization="provenanceaudiooutput_dashinit.mp4" timescale="48000" startNumber="1" duration="96000"/>
-   <Representation id="1" mimeType="audio/mp4" codecs="mp4a.40.2" audioSamplingRate="48000" bandwidth="128712">
-    <AudioChannelConfiguration schemeIdUri="urn:mpeg:dash:23003:3:audio_channel_configuration:2011" value="2"/>
-   </Representation>
-  </AdaptationSet>
-  <AdaptationSet segmentAlignment="true" maxWidth="320" maxHeight="240" maxFrameRate="30000/1001" par="4:3" lang="eng" startWithSAP="1">
-   <SegmentTemplate media="provenancevideooutput_dash$Number$.m4s" initialization="provenancevideooutput_dashinit.mp4" timescale="30000" startNumber="1" duration="60000"/>
-   <Representation id="2" mimeType="video/mp4" codecs="avc1.42400D" width="320" height="240" frameRate="30000/1001" sar="1:1" bandwidth="479072">
-   </Representation>
-  </AdaptationSet>
- </Period>
-</MPD>
