--- samples/microsoft/video_samples/DASH/audio_video_sample/provenanceaudiooutput_dash1_dump.xml
@@ -1,121 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ISOBaseMediaFileTrace>
-<!--MP4Box dump trace-->
-<IsoMediaFile xmlns="urn:mpeg:isobmff:schema:file:2016" Name="provenanceaudiooutput_dash1.m4s">
-<SegmentTypeBox Size="24" Type="styp" Specification="p12" Container="file" MajorBrand="msdh" MinorVersion="0">
-<BrandEntry AlternateBrand="msdh"/>
-<BrandEntry AlternateBrand="msix"/>
-</SegmentTypeBox>
-<SegmentIndexBox Size="44" Type="sidx" Version="0" Flags="0" Specification="p12" Container="file" reference_ID="1" timescale="48000" earliest_presentation_time="0" first_offset="0">
-<Reference type="0" size="16050" duration="95359" startsWithSAP="1" SAP_type="1" SAPDeltaTime="0"/>
-</SegmentIndexBox>
-<UUIDBox Size="141" Type="uuid" UUID="{D8FEC3D6-1B0E483C-92975828-877EC481}" Specification="unknown" Container="unknown" >
-</UUIDBox>
-<MovieFragmentBox Size="804" Type="moof" Specification="p12" Container="file" TrackFragments="1">
-<MovieFragmentHeaderBox Size="16" Type="mfhd" Version="0" Flags="0" Specification="p12" Container="moof" FragmentSequenceNumber="1">
-</MovieFragmentHeaderBox>
-<TrackFragmentBox Size="780" Type="traf" Specification="p12" Container="moof" >
-<TrackFragmentHeaderBox Size="16" Type="tfhd" Version="0" Flags="131072" Specification="p12" Container="traf" TrackID="1" BaseDataOffset="moof">
-</TrackFragmentHeaderBox>
-<TrackFragmentBaseMediaDecodeTimeBox Size="16" Type="tfdt" Version="0" Flags="0" Specification="p12" Container="traf" baseMediaDecodeTime="0">
-</TrackFragmentBaseMediaDecodeTimeBox>
-<TrackRunBox Size="740" Type="trun" Version="0" Flags="769" Specification="p12" Container="traf" SampleCount="90" DataOffset="812">
-<TrackRunEntry Duration="4223" Size="160"/>
-<TrackRunEntry Duration="1024" Size="171"/>
-<TrackRunEntry Duration="1024" Size="136"/>
-<TrackRunEntry Duration="1024" Size="158"/>
-<TrackRunEntry Duration="1024" Size="148"/>
-<TrackRunEntry Duration="1024" Size="149"/>
-<TrackRunEntry Duration="1024" Size="164"/>
-<TrackRunEntry Duration="1024" Size="140"/>
-<TrackRunEntry Duration="1024" Size="160"/>
-<TrackRunEntry Duration="1024" Size="155"/>
-<TrackRunEntry Duration="1024" Size="164"/>
-<TrackRunEntry Duration="1024" Size="165"/>
-<TrackRunEntry Duration="1024" Size="171"/>
-<TrackRunEntry Duration="1024" Size="157"/>
-<TrackRunEntry Duration="1024" Size="167"/>
-<TrackRunEntry Duration="1024" Size="159"/>
-<TrackRunEntry Duration="1024" Size="181"/>
-<TrackRunEntry Duration="1024" Size="165"/>
-<TrackRunEntry Duration="1024" Size="164"/>
-<TrackRunEntry Duration="1024" Size="184"/>
-<TrackRunEntry Duration="1024" Size="179"/>
-<TrackRunEntry Duration="1024" Size="169"/>
-<TrackRunEntry Duration="1024" Size="169"/>
-<TrackRunEntry Duration="1024" Size="187"/>
-<TrackRunEntry Duration="1024" Size="172"/>
-<TrackRunEntry Duration="1024" Size="175"/>
-<TrackRunEntry Duration="1024" Size="168"/>
-<TrackRunEntry Duration="1024" Size="173"/>
-<TrackRunEntry Duration="1024" Size="173"/>
-<TrackRunEntry Duration="1024" Size="166"/>
-<TrackRunEntry Duration="1024" Size="190"/>
-<TrackRunEntry Duration="1024" Size="190"/>
-<TrackRunEntry Duration="1024" Size="177"/>
-<TrackRunEntry Duration="1024" Size="166"/>
-<TrackRunEntry Duration="1024" Size="192"/>
-<TrackRunEntry Duration="1024" Size="213"/>
-<TrackRunEntry Duration="1024" Size="184"/>
-<TrackRunEntry Duration="1024" Size="157"/>
-<TrackRunEntry Duration="1024" Size="148"/>
-<TrackRunEntry Duration="1024" Size="153"/>
-<TrackRunEntry Duration="1024" Size="159"/>
-<TrackRunEntry Duration="1024" Size="149"/>
-<TrackRunEntry Duration="1024" Size="163"/>
-<TrackRunEntry Duration="1024" Size="171"/>
-<TrackRunEntry Duration="1024" Size="176"/>
-<TrackRunEntry Duration="1024" Size="183"/>
-<TrackRunEntry Duration="1024" Size="193"/>
-<TrackRunEntry Duration="1024" Size="169"/>
-<TrackRunEntry Duration="1024" Size="162"/>
-<TrackRunEntry Duration="1024" Size="167"/>
-<TrackRunEntry Duration="1024" Size="147"/>
-<TrackRunEntry Duration="1024" Size="162"/>
-<TrackRunEntry Duration="1024" Size="167"/>
-<TrackRunEntry Duration="1024" Size="161"/>
-<TrackRunEntry Duration="1024" Size="160"/>
-<TrackRunEntry Duration="1024" Size="157"/>
-<TrackRunEntry Duration="1024" Size="193"/>
-<TrackRunEntry Duration="1024" Size="166"/>
-<TrackRunEntry Duration="1024" Size="186"/>
-<TrackRunEntry Duration="1024" Size="168"/>
-<TrackRunEntry Duration="1024" Size="177"/>
-<TrackRunEntry Duration="1024" Size="176"/>
-<TrackRunEntry Duration="1024" Size="185"/>
-<TrackRunEntry Duration="1024" Size="173"/>
-<TrackRunEntry Duration="1024" Size="165"/>
-<TrackRunEntry Duration="1024" Size="164"/>
-<TrackRunEntry Duration="1024" Size="157"/>
-<TrackRunEntry Duration="1024" Size="158"/>
-<TrackRunEntry Duration="1024" Size="186"/>
-<TrackRunEntry Duration="1024" Size="169"/>
-<TrackRunEntry Duration="1024" Size="178"/>
-<TrackRunEntry Duration="1024" Size="173"/>
-<TrackRunEntry Duration="1024" Size="176"/>
-<TrackRunEntry Duration="1024" Size="172"/>
-<TrackRunEntry Duration="1024" Size="177"/>
-<TrackRunEntry Duration="1024" Size="157"/>
-<TrackRunEntry Duration="1024" Size="181"/>
-<TrackRunEntry Duration="1024" Size="175"/>
-<TrackRunEntry Duration="1024" Size="165"/>
-<TrackRunEntry Duration="1024" Size="168"/>
-<TrackRunEntry Duration="1024" Size="174"/>
-<TrackRunEntry Duration="1024" Size="170"/>
-<TrackRunEntry Duration="1024" Size="178"/>
-<TrackRunEntry Duration="1024" Size="213"/>
-<TrackRunEntry Duration="1024" Size="149"/>
-<TrackRunEntry Duration="1024" Size="152"/>
-<TrackRunEntry Duration="1024" Size="187"/>
-<TrackRunEntry Duration="1024" Size="170"/>
-<TrackRunEntry Duration="1024" Size="170"/>
-<TrackRunEntry Duration="1024" Size="165"/>
-</TrackRunBox>
-</TrackFragmentBox>
-</MovieFragmentBox>
-<MediaDataBox Size="15246" Type="mdat" Specification="p12" Container="file" dataSize="15238">
-</MediaDataBox>
-</IsoMediaFile>
-<Tracks>
-</Tracks>
-</ISOBaseMediaFileTrace>
