--- samples/microsoft/video_samples/DASH/audio_video_sample/video_prog_index.m3u8
@@ -1,29 +1,0 @@
-#EXTM3U
-#EXT-X-TARGETDURATION:2
-#EXT-X-VERSION:7
-#EXT-X-MEDIA-SEQUENCE:1
-#EXT-X-PLAYLIST-TYPE:VOD
-#EXT-X-INDEPENDENT-SEGMENTS
-#EXT-X-MAP:URI="provenancevideooutput_dashinit.mp4"
-#EXTINF:2,
-provenancevideooutput_dash1.m4s
-#EXTINF:2,
-provenancevideooutput_dash2.m4s
-#EXTINF:2,
-provenancevideooutput_dash3.m4s
-#EXTINF:2,
-provenancevideooutput_dash4.m4s
-#EXTINF:2,
-provenancevideooutput_dash5.m4s
-#EXTINF:2,
-provenancevideooutput_dash6.m4s
-#EXTINF:2,
-provenancevideooutput_dash7.m4s
-#EXTINF:2,
-provenancevideooutput_dash8.m4s
-#EXTINF:2
-provenancevideooutput_dash9.m4s
-#EXTINF:2,
-provenancevideooutput_dash10.m4s
-#EXTINF:2,
-provenancevideooutput_dash11.m4s
